Move sequence steps and inputs to sub components.
